Allowed to build Poco library with our compiler options [#METR-19593].

+++ libs/libpoco/Foundation/src/HashStatistic.cpp
@@ -21,10 +21,10 @@ namespace Poco {
 
 
 HashStatistic::HashStatistic(
-	UInt32 tableSize, 
-	UInt32 numEntries, 
-	UInt32 numZeroEntries, 
-	UInt32 maxEntry, 
+	UInt32 tableSize,
+	UInt32 numEntries,
+	UInt32 numZeroEntries,
+	UInt32 maxEntry,
 	std::vector<UInt32> details):
 	_sizeOfTable(tableSize),
 	_numberOfEntries(numEntries),
@@ -48,7 +48,7 @@ std::string HashStatistic::toString() const
 	str << "  MaxEntry: " << _maxEntriesPerHash << "\n";
 	str << "  AvgEntry: " << avgEntriesPerHash() << ", excl Zero slots: " << avgEntriesPerHashExclZeroEntries() << "\n";
 	str << "  DetailedStatistics: \n";
-	for (int i = 0; i < _detailedEntriesPerHash.size(); ++i)
+	for (size_t i = 0; i < _detailedEntriesPerHash.size(); ++i)
 	{
 		// 10 entries per line
 		if (i % 10 == 0)
