+++ libs/libpoco/JSON/src/Parser.cpp
@@ -185,7 +185,7 @@ bool Parser::push(int mode)
 	_top += 1;
 	if (_depth < 0)
 	{
-		if (_top >= _stack.size())
+		if ((size_t)_top >= _stack.size())
 			_stack.resize(_stack.size() * 2, true);
 	}
 	else
@@ -312,7 +312,7 @@ Parser::CharIntType Parser::decodeUnicodeChar()
 		uc |= x << i;
 	}
 
-	if ( !_allowNullByte && uc == 0 ) return 0; 
+	if ( !_allowNullByte && uc == 0 ) return 0;
 
 	// clear UTF-16 char from buffer
 	_parseBuffer.resize(_parseBuffer.size() - 6);
@@ -348,7 +348,7 @@ Parser::CharIntType Parser::decodeUnicodeChar()
 		}
 		else if (isLowSurrogate(uc))
 		{
-			// low surrogate without a preceding high surrogate 
+			// low surrogate without a preceding high surrogate
 			return 0;
 		}
 		else
@@ -396,7 +396,7 @@ void Parser::parseBuffer()
 					break;
 				}
 			case JSON_T_FLOAT:
-				{ 
+				{
 					// Float can't end with a dot
 					if (_parseBuffer[_parseBuffer.size() - 1] == '.' ) throw SyntaxException("JSON syntax error");
 
@@ -469,34 +469,34 @@ int Parser::utf8CheckFirst(char byte)
 	if(u < 0x80)
 		return 1;
 
-	if (0x80 <= u && u <= 0xBF) 
+	if (0x80 <= u && u <= 0xBF)
 	{
 		// second, third or fourth byte of a multi-byte
 		// sequence, i.e. a "continuation byte"
 		return 0;
 	}
-	else if(u == 0xC0 || u == 0xC1) 
+	else if(u == 0xC0 || u == 0xC1)
 	{
 		// overlong encoding of an ASCII byte
 		return 0;
 	}
-	else if(0xC2 <= u && u <= 0xDF) 
+	else if(0xC2 <= u && u <= 0xDF)
 	{
 		// 2-byte sequence
 		return 2;
 	}
-	else if(0xE0 <= u && u <= 0xEF) 
+	else if(0xE0 <= u && u <= 0xEF)
 	{
 		// 3-byte sequence
 		return 3;
 	}
-	else if(0xF0 <= u && u <= 0xF4) 
+	else if(0xF0 <= u && u <= 0xF4)
 	{
 		// 4-byte sequence
 		return 4;
 	}
-	else 
-	{ 
+	else
+	{
 		// u >= 0xF5
 		// Restricted (start of 4-, 5- or 6-byte sequence) or invalid UTF-8
 		return 0;
