+++ libs/libpoco/Net/src/ICMPEventArgs.cpp
@@ -33,11 +33,11 @@ namespace Net {
 
 
 ICMPEventArgs::ICMPEventArgs(const SocketAddress& address, int repetitions, int dataSize, int ttl):
-	_address(address), 
+	_address(address),
 	_sent(0),
-	_dataSize(dataSize), 
-	_ttl(ttl), 
-	_rtt(repetitions, 0), 
+	_dataSize(dataSize),
+	_ttl(ttl),
+	_rtt(repetitions, 0),
 	_errors(repetitions)
 {
 }
@@ -54,10 +54,10 @@ std::string ICMPEventArgs::hostName() const
 	{
 		return DNS::resolve(_address.host().toString()).name();
 	}
-	catch (HostNotFoundException&) 
+	catch (HostNotFoundException&)
 	{
 	}
-	catch (NoAddressFoundException&) 
+	catch (NoAddressFoundException&)
 	{
 	}
 	catch (DNSException&)
@@ -103,7 +103,7 @@ int ICMPEventArgs::received() const
 {
 	int received = 0;
 
-	for (int i = 0; i < _rtt.size(); ++i) 
+	for (unsigned i = 0; i < _rtt.size(); ++i)
 	{
 		if (_rtt[i]) ++received;
 	}
@@ -113,7 +113,7 @@ int ICMPEventArgs::received() const
 
 void ICMPEventArgs::setError(int index, const std::string& text)
 {
-	if (index >= _errors.size()) 
+	if ((size_t)index >= _errors.size())
 		throw InvalidArgumentException("Supplied index exceeds vector capacity.");
 
 	_errors[index] = text;
@@ -122,7 +122,7 @@ void ICMPEventArgs::setError(int index, const std::string& text)
 
 const std::string& ICMPEventArgs::error(int index) const
 {
-	if (0 == _errors.size()) 
+	if (0 == _errors.size())
 		throw InvalidArgumentException("Supplied index exceeds vector capacity.");
 
 	if (-1 == index) index = _sent - 1;
@@ -133,7 +133,7 @@ const std::string& ICMPEventArgs::error(int index) const
 
 void ICMPEventArgs::setReplyTime(int index, int time)
 {
-	if (index >= _rtt.size()) 
+	if ((size_t)index >= _rtt.size())
 		throw InvalidArgumentException("Supplied index exceeds array capacity.");
 	if (0 == time) time = 1;
 	_rtt[index] = time;
@@ -142,7 +142,7 @@ void ICMPEventArgs::setReplyTime(int index, int time)
 
 int ICMPEventArgs::replyTime(int index) const
 {
-	if (0 == _rtt.size()) 
+	if (0 == _rtt.size())
 		throw InvalidArgumentException("Supplied index exceeds array capacity.");
 
 	if (-1 == index) index = _sent - 1;
@@ -154,7 +154,7 @@ int ICMPEventArgs::replyTime(int index) const
 int ICMPEventArgs::avgRTT() const
 {
 	if (0 == _rtt.size()) return 0;
-	
+
 	return (int) (std::accumulate(_rtt.begin(), _rtt.end(), 0) / _rtt.size());
 }
